Fixed #21, Default Recording State
The default value for setting the recording state was invalid and does
not exist in Matterhorn. As it was never used and you would usually want
to set a specific state anyway, this commit removes the default value of
that method completely.

Thanks to Spoil001 {1] for noticing and letting me know.

[1] https://github.com/Spoil001

Signed-off-by: Lars Kiesow <lkiesow@uos.de>