Use the ServiceRegistry to Determine Services
Matterhorn allows it to scatter services across several servers, meaning
that there is no base URL to use for all service REST endpoints. But all
services are registered at the service registry from which we can
retrieve the exact locations.

This also allows us to use several services of one type. This is
interesting in particular when thinking about ingest services.

To ensure all available ingest services are used equally, the service to
use for a particular ingest is chosen at random.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>