User Interface Update
This patch updates the user interface to make use of the new information
we have due to the database.

The new interface now shows basic statistics as well as a list of
processed and upcoming recordings.

It also updates the design to look more like something related to
Opencast. It vaguely resembles the Opencast admin interface as well as
(or maybe more so) the Opencast REST documentation.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>