Simplify pyCA Shutdown (#75)
This patch introduces a global shutdown method instead of using the
termination variables for every module. That way, it is easier to shut
pyCA down from anywhere in the code.

It also makes testing easier since we can now determine if code should
run through before the termination is triggered instead if having to
assign the value of the termination variable in advance.