PyCA 2.0
This is the release of pyCA 2.0.

Version 2.0 is a major redesign of the source code and a significant
step forward for more robustness, flexibility and usability.

Main Features
-------------

- Introduction of new worker concept, enabling the capture agent to
  perform parallel capture, ingest and scheduling tasks. It also makes
  it possible to control and launch parts of the capture agent
  separately and ensures that potential errors are well contained and do
  not affect all of the capture agent.

- New user interface to see what is going on on the capture agent. The
  web interface provides information about current and past recordings
  as well as the currently buffered schedule. The interface is optimized
  for a mobile view as well.

- Automatic verification of the configuration to prevent issues early.
  This includes type and (in some cases) value checks for configuration
  options.

- 96% unit test coverage for the code to ensure that things actually
  work and new patches do not accidentally break known behavior.

- Python 2 and Python 3 compatibility to ensure that pyCA can be run on
  all types of devices and operating systems.

- Fixes for a lot of minor issues like general Unicode compatibility,
  user feedback, system logging, …

New Release Process
-------------------

PyCA will now follow a more agile release process with smaller, but more
frequent releases. PyCA 2.1 is already planned and will contain features
like a REST API, a more responsive web interface and capture process
control mechanisms.

Credits
-------

A special thanks to Jan Koppe and Sven Haardiek which helped a lot to
make this release possible by submitting and reviewing code as well as
complaining about potential issues. Thanks!

Signed-off-by: Lars Kiesow <lkiesow@uos.de>