Strip / from end of base url
This patch removes one `/` from the end of the configured base URL in
pyCA since it is very likely a configuration issue. PyCA warns warns
about the user about this action so he can properly fix the
configuration or, in the unlikely case that the URL really should end
with `/`, intervene and add a second `/` to the configured value.

This fixes #132