Fix display of next schedule
This patch fixes the logging of the next schedule. Previously, the next
in line database entry of all upcoming events would be logged. This was
not guaranteed to be the next scheduled recording since the events in
the database were not guaranteed to be ordered by start date.

This patch fixes the issue by selecting only the event with the next
start date from the database.