add: wait for a randomized delay before ingest
when many pyCA instances try to ingest large recordings at once, the
ingest server might get overloaded, resulting in timeouts and failed
ingests.

this patch allows the user to configure a randomized time range (in
seconds, min and max) that pyCA will wait before trying to ingest a
finished recording, which will spread out the load on the ingest server.