Update Agent State Properly
pyCA's status is set by the agent state service which is not guaranteed
to start or stop before or after any of the other processes upon which
the current status is based. This may lead to race conditions when
starting or stopping pyCA in which cases:

- pyCA may appear as offline after it is started
- pyCA may not be set to offline after it is shut down

While this issue does not have any negative effect on recording
capabilities (pyCA will still record even if the state is still set to
offline), this may cause some confusion.

This patch fixes the problem by setting the state immediately after
relevant services are started and when they are shut down.