New UI reload
This patch changes the ui from reloading itself to using Vue.js and only
reloading the specific data.

It completely removed flask templating and slightly changes and expand the JSON
API.
The service API is now splitted in `upcoming` and `recorded` and there are new
endpoints for getting the name of the capture agent and a list of preview
images.

It also adds NPM and [Parcel.js](https://parceljs.org) for managing
the JavaScript dependencies and building a static version of the UI for
production.

There are also some slightly changes in the features.
The refresh is now a url parameter and is received from the backend on redirect
from `/` to the `index.html`.