Service Configuration and Management
This patch overhauls the Systemd services. It adds a master service to
allow for easier management of pyCA, starting and stopping it using one
easy command. It also changes the main configuration location to allow
for multiple configuration files.

The master service controls all other services, allowing to use the
following commands to shut them down as well:

    systemctl start pyca.service
    systemctl stop pyca.service
    systemctl restart pyca.service

Not that for this to work, the other services must be anabled. This also
means that you can still exclude services like the user interface by
disabling them. Usually, though, you want to run this to ensure pyCA is
fully functional as service:

    systemctl enable pyca-agentstate.service
    systemctl enable pyca-capture.service
    systemctl enable pyca-ingest.service
    systemctl enable pyca-schedule.service
    systemctl enable pyca-ui.service

The main configuration file is now located at `/etc/pyca/pyca.conf` The
additional sirectory ensures that additional configuration files can be
stored. One new default configuration is `/etc/pyca/gunicorn.conf.py`
which is included in the Systemd unit file for the user interface and
for which a default one is included in the `etc` folder.