Logging API Alternative
This patch is a draft for an alternative to storing all logs in the
database (#213). Instead, this patch allows to specify an arbitrary
command responsible for retrieving the logs.

While this gives us less control over the logs (e.g. it is hard to
distinguish different log level), this makes the code much simpler and
removes the burden, the logging puts on the database and the problem
that the database grows indefinitely.