Database Initialization
When all services are started simultaneously with no existing database,
it may happen that multiple services try creating the database structure
and some of them fail which causes the entire service to fail.

This patch ensures that the database structure is created before
starting the services removing the problem when starting all services
with pyCA's main method.

The problem can still appear when services are handled separately but
that is far less likely and there are usually safeguards in place – e.g.
systemd will restart the failed service which will automatically fix
the problem as well.