Update Recording Table Width
This patch updates the column width of the main recordings table
preventing actions to be shown more or less in the status label. The
additional space is taken from the title column which maximum width is
constricted further.

To make up for this, the complete title is now contained in a title tag
which is shown when you hover over the title cells of the table.