Update pyca.conf
New entry for Input names in pyca.conf.  Each Input belongs to one flavor and to one output file, so Input names, flavor names and output names must be unique and must have the same number. The inputs are sent to the opencast-server and should/must be selected with new schedule entries.
If inputs is not defined (default is=''), each track/flavor/output-file is added to the media package and uploaded. Flavor names must be unique because they are used to find the input behind the output file. This is a work-around because I did not want to change the database entries. It would be better to have the input name in the event.track database entry as [input, flavor, output_file], but then more changes are needed.