Don't run tests on Dependabot twice
Tests run twice on Dependabot pull requests since they react to both the
`push` and the `pull_request` action. While we want to keep the push
action in general, it doesn't really help in case of Dependabot.

This patch prevents running on push actions on branches starting with
`dependabot`. This should be an easy fix preventing the issue with a low
likelihood of effecting anything else.