Add "inputs" support to pyCA
Adds support for sending configuration to the agent
configuration endpoint in Opencast.
Adds a new config option "inputs". It allows users to specifiy for a
given event which tracks they would like to show up in the final
recording. For example, if the agent is capable of recording both
"presenter/source" and "presentation/source", a user may
select only "presentation/source" as input.